ENH: set default number of threads to 1

+++ src/esinfo/envinfo.h
@@ -2,16 +2,14 @@
 #ifndef SRC_ESINFO_ENVINFO_H_
 #define SRC_ESINFO_ENVINFO_H_
 
-#include "mpi.h"
-#include <cstddef>
-
 namespace espreso {
 namespace info {
 namespace env {
-	extern size_t MKL_NUM_THREADS;
-	extern size_t OMP_NUM_THREADS;
-	extern size_t SOLVER_NUM_THREADS;
-	extern size_t PAR_NUM_THREADS;
+
+	extern int MKL_NUM_THREADS;
+	extern int OMP_NUM_THREADS;
+	extern int SOLVER_NUM_THREADS;
+	extern int PAR_NUM_THREADS;
 
 	void set();
 }
